refactor: move src/* to lib/
